[BIN] Move into one daemon
Seednodes are now run with the --seednode and --seednode-ip options rather than a seperate bin
